tests/wpt/web-platform-tests/css/css-align/parsing/place-items-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: place-items sets longhands</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-place-items">
<meta name="assert" content="place-items supports the full grammar '<align-items> <justify-items>?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/shorthand-testcommon.js"></script>
</head>
<body>
<script>
test_shorthand_value('place-items', 'normal', {
  'align-items': 'normal',
  'justify-items': 'normal'
});

test_shorthand_value('place-items', 'first baseline', {
  'align-items': 'baseline',
  'justify-items': 'baseline'
});

test_shorthand_value('place-items', 'last baseline flex-start', {
  'align-items': 'last baseline',
  'justify-items': 'flex-start'
});

test_shorthand_value('place-items', 'stretch right legacy', {
  'align-items': 'stretch',
  'justify-items': 'legacy right'
});
</script>
</body>
</html>
